+++ ui/components/MesheryApplications.js
@@ -15,10 +15,10 @@ import { UnControlled as CodeMirror } from "react-codemirror2";
 import DeleteIcon from "@material-ui/icons/Delete";
 import SaveIcon from '@material-ui/icons/Save';
 import UploadIcon from "@material-ui/icons/Publish";
-import PromptComponent from "./PromptComponent";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import MUIDataTable from "mui-datatables";
+import PromptComponent from "./PromptComponent";
 import Moment from "react-moment";
 import { withSnackbar } from "notistack";
 import CloseIcon from "@material-ui/icons/Close";
@@ -102,8 +102,8 @@ function MesheryApplications({
   const [search] = useState("");
   const [sortOrder] = useState("");
   const [count, setCount] = useState(0);
-  const [pageSize, setPageSize] = useState(10);
   const modalRef = useRef(null);
+  const [pageSize, setPageSize] = useState(10);
   const [applications, setApplications] = useState([]);
   const [selectedRowData, setSelectedRowData] = useState(null);
   const DEPLOY_URL = '/api/application/deploy';
@@ -373,16 +373,16 @@ function MesheryApplications({
     }
   });
 
-  async function deleteApplication(id) {
-    let response = await modalRef.current.show({
-      title : "Delete Application?",
+  async function showModal() {
+    let response = await modalRef.current.show({ title : "Delete Aplication?",
 
       subtitle : "Are you sure you want to delete this application?",
 
-      options : ["yes", "no"],
+      options : ["yes", "no"], })
+    return response;
+  }
 
-    })
-    if (response === "NO") return
+  async function deleteApplication(id) {
     dataFetch(
       `/api/application/${id}`,
       {
@@ -417,7 +417,6 @@ function MesheryApplications({
     responsive : "scrollFullHeight",
     resizableColumns : true,
     serverSide : true,
-    selection : true,
     count,
     rowsPerPage : pageSize,
     rowsPerPageOptions : [10, 20, 25],
@@ -427,9 +426,15 @@ function MesheryApplications({
     download : false,
     customToolbar : CustomToolbar(uploadHandler),
 
-    onRowsDelete : function handleDelete(row) {
-      const fid = Object.keys(row.lookup).map(idx => applications[idx]?.id)
-      fid.forEach(fid => deleteApplication(fid))
+    onRowsDelete : async function handleDelete(row) {
+      let response = await showModal()
+      console.log(response)
+      if (response === "yes") {
+        const fid = Object.keys(row.lookup).map(idx => applications[idx]?.id)
+        fid.forEach(fid => deleteApplication(fid))
+      }
+      if (response === "no")
+        fetchApplications(page, pageSize, search, sortOrder);
     },
 
     onTableChange : (action, tableState) => {
